Added watson backup, some refactors
